requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
azure-ai-formrecognizer==3.3.2
python-decouple==3.8
